# Conflicts:
#	app/src/main/java/com/mnassa/extensions/PostExtensions.kt
#	app/src/main/java/com/mnassa/screen/posts/need/create/CreateNeedViewModel.kt
#	app/src/main/java/com/mnassa/screen/posts/need/create/CreateNeedViewModelImpl.kt
#	app/src/main/java/com/mnassa/screen/posts/need/details/NeedDetailsViewModelImpl.kt
#	data/src/main/java/com/mnassa/data/repository/DatabaseContract.kt
#	data/src/main/res/values/android_dictionary.xml
#	domain/src/main/java/com/mnassa/domain/model/PostModel.kt

+++ domain/src/main/java/com/mnassa/domain/repository/WalletRepository.kt
@@ -12,7 +12,7 @@ interface WalletRepository {
     suspend fun getSpentPointsCount(): ReceiveChannel<Long>
     suspend fun getGainedPointsCount(): ReceiveChannel<Long>
     suspend fun getTransactions(): ReceiveChannel<List<TransactionModel>>
-    suspend fun getDefaultRewardingPoints(): Int
+    suspend fun getDefaultRewardingPoints(): Long
 
     suspend fun sendPoints(amount: Long, recipientId: String, description: String?)
     suspend fun sendPointsForComment(rewardModel: RewardModel)
